completely disable rm env syncing
This allows artificially small env size.
Also this avoids the problem of overwriting current PATH
with the one from config.sys, which is usually "PATH=." to
work around freedos bugs. It was copied to new env segment,
left there because syncing was disabled, and later copied
to PM, overwriting current PATH. But now its not even copied
to the new sigment, so it can't do any harm.
Tail (argv0) is also not copied.
Hope djgpp runtime makes a back-up.