propagate COMSPEC to env seg [fixes #37]
Commit 9ccf649 applies the optimization to not sync RM env
with PM one. Every DOS process has its own environment copy,
so for comcom32 it is not needed.
It appears that WordPerfect looks up the COMSPEC var in a
parent segment, rather than in its own. To allow it to work,
we need to put COMSPEC into our env segment.