del: use findnext() to avoid leaks [fixes #84]
It appears findnext() needs to be called until the search is
completed, and only in that case djgpp closes the findfirst handle.
Call findnext() before removing the file to stay on a safe side.
Perhaps glob() would be a better solution.

Using only findfirst() was fine before LFNs were enabled, because
the SFN's findfirst() doesn't need to close searches. But with
LFNs this eventually overflows the buffer.