fix lfn ff handle leak
deltree must restart ff on backtrack in SFN mode - see 7e3365bd.
In LFN mode there are ff handles that require findnext() until
completion, or they are leaked. So close them explicitly.

Regression of 7e3365bd - LFNs were not enabled back then.

Note: it is possible to special-case LFN mode and not restart
ff, relying on open handles. But this means coding difference
with SFN mode, and also there may be insufficient ff handles
in a redirector to traverse the large dir tree.