implement sequential read mode [closes #108]
This is much faster than open/close bat file per every line.
Because after every open, we also did fgets() multiple times
to reach the next line.

In sequential mode, the bat file is closed only when its processing
is done, which may be exit, executing another bat w/o call, or an error.
Multiple reads are done in this mode only when searching for the goto label.